endpoint: move deletion deadlock unit test to `pkg/endpoint`
Signed-off by: Ian Vernon <ian@cilium.io>
