+++ pkg/datapath/fake/datapath.go
@@ -15,14 +15,17 @@
 package fake
 
 import (
+	"context"
 	"io"
 
 	"github.com/cilium/cilium/pkg/datapath"
+	"github.com/cilium/cilium/pkg/datapath/loader/metrics"
 )
 
 type fakeDatapath struct {
 	node           datapath.NodeHandler
 	nodeAddressing datapath.NodeAddressing
+	loader         datapath.Loader
 }
 
 // NewDatapath returns a new fake datapath
@@ -30,6 +33,7 @@ func NewDatapath() datapath.Datapath {
 	return &fakeDatapath{
 		node:           NewNodeHandler(),
 		nodeAddressing: NewNodeAddressing(),
+		loader:         &fakeLoader{},
 	}
 }
 
@@ -77,5 +81,40 @@ func (f *fakeDatapath) SupportsOriginalSourceAddr() bool {
 }
 
 func (f *fakeDatapath) Loader() datapath.Loader {
-	return nil
+	return f.loader
+}
+
+// Loader is an interface to abstract out loading of datapath programs.
+type fakeLoader struct {
+}
+
+func (f *fakeLoader) CompileAndLoad(ctx context.Context, ep datapath.Endpoint, stats *metrics.SpanStat) error {
+	panic("implement me")
+}
+
+func (f *fakeLoader) CompileOrLoad(ctx context.Context, ep datapath.Endpoint, stats *metrics.SpanStat) error {
+	panic("implement me")
+}
+
+func (f *fakeLoader) ReloadDatapath(ctx context.Context, ep datapath.Endpoint, stats *metrics.SpanStat) error {
+	panic("implement me")
+}
+
+func (f *fakeLoader) EndpointHash(cfg datapath.EndpointConfiguration) (string, error) {
+	panic("implement me")
+}
+
+func (f *fakeLoader) DeleteDatapath(ctx context.Context, ifName, direction string) error {
+	panic("implement me")
+}
+
+func (f *fakeLoader) Unload(ep datapath.Endpoint) {
+}
+
+func (f *fakeLoader) Init(d datapath.ConfigWriter, nodeCfg *datapath.LocalNodeConfiguration) {
+	panic("implement me")
+}
+
+func (f *fakeLoader) CallsMapPath(id uint16) string {
+	return ""
 }
